Add files via upload
This is an advanced message/sms/email spam classifier which is created by using the dataset that was provided by UCi machine learning repository. The classifier is tested on several machine learning algorithms to get the best accuracy and precision scores. In the end the best model was chosen and converted into a webapp that can be found on my replit account.