Add files via upload
This is a DeepFace model that recognizes the face of the person uploaded and then is compares with the celebrities dataset and gives an outcome according to the similarity distances calculated. It gives a prediction according to its calculations that which celebrity do you look like. You can train it according to your own datasets and images.