Add files via upload
This notebook contains an image captioning project full code script that is ready to use. It contains a pre trained VGG16 model that is trained on a very fine image data set that i downloaded from kaggle. All preprocessing and feature engineering is included. The model provides a very fine accuracy. You can download the code and modify according to your recommendation.